+++ src/lib/is-browser.ts
@@ -33,8 +33,8 @@ const isStandardBrowserEnv = () => {
 const isWebWorkerEnv = () =>
 	Boolean(
 		typeof self === 'object' &&
-			self?.constructor?.name?.includes('WorkerGlobalScope') &&
-			typeof Deno === 'undefined',
+		self?.constructor?.name?.includes('WorkerGlobalScope') &&
+		typeof Deno === 'undefined',
 	)
 
 const isReactNativeEnv = () =>